feat: User 정보 등록 API
요청에 Access Token을 받기 때문에 Security에서 경로 제외함

+++ src/main/java/zip/ootd/ootdzip/user/data/UserRegisterReq.java
@@ -18,11 +18,11 @@ public class UserRegisterReq {
     @JsonProperty(required = true)
     private LocalDate birthdate;
 
-    private Long height;
+    private Integer height;
 
     private Boolean showHeight;
 
-    private Long weight;
+    private Integer weight;
 
     private Boolean showWeight;
 }
